+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Update dependencies
+- Add `BoxError` to `InvalidRequestor` and `InvalidReactor` errors
 
 ## 6.1.0 - 2021-10-11
 - Update dependencies